==> notebooks/Kervy/04-Week_Model.ipynb <==
%load_ext autoreload
%autoreload 2
%matplotlib inline
%config InlineBackend.figure_format = 'retina'

from pandas_summary import DataFrameSummary
from pathlib import Path

import os, math
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
import seaborn as sns

pd.set_option('display.max_columns', 1000)
pd.set_option('display.max_rows', 400)
sns.set()
os.chdir('/home/krivas/projects/analysis-project/')
from src.utils import convert_int, convert_date, add_date_diffs, add_days_count, add_date_features,\
                        make_set, run_model, apply_cats, get_week_labels


DATA        = Path('data')
RAW         = DATA/'raw'
PROCESSED   = DATA/'processed'

# Reading data

csv = pd.read_csv(PROCESSED/'HistoricoJob.csv')

csv.columns

csv[['Fecha_Carga_Scheduler', 'Fecha_Ejec_Inicio']] 

target_date = 20180515
n_weeks = 10

week_labels = get_week_labels(n_weeks, 2018, 5, 14)
week_labels

%%time
weeks = [make_set(csv[csv.Fecha_Ejec_Inicio_Int.apply(lambda x: w1<=x<w2)], 
                    csv.loc[csv['Fecha_Ejec_Inicio_Int'] == target_date]) for w1, w2 in zip(week_labels[:-1], week_labels[1:])]

weeks.append(make_set(csv.loc[csv['Fecha_Ejec_Inicio_Int'] == target_date], csv.loc[csv['Fecha_Ejec_Inicio_Int'] == target_date]))

# Preprocessing

%%time
for temp in weeks:
    add_datediffs(temp, csv)
    add_dayscount(temp, csv)
    add_datefeatures(temp)

weeks[1].head()

cat_cols = ['DiaSemana']

%%time
for w in weeks:
    for col in cat_cols:
        w[col] = w[col].astype('category').cat.codes

# Model

#from src.utils import run_model
from lightgbm import LGBMRegressor

target_col = 'duracion_int'
pred_col = 'Prediccion'
id_cols = ['Fecha_Ejec_Inicio_Int', 'Id_Job', 'Id_Malla']
metric_names = ['mse']

weeks[len(week_labels) - 1][pred_col] = 0

%%time
results = None
output = []
for i, w in enumerate(weeks[1:]):
    train, val, test = pd.concat(weeks[:i+1], axis=0), w, weeks[-1]
    X_train, y_train = train.drop(id_cols + [target_col], axis=1), \
                       train[target_col]
    if pred_col in val.columns: # when test acts as validation
        X_val, y_val = None, None
    else:
        X_val, y_val = val.drop(id_cols + [target_col], axis=1), \
                       val[target_col]
    X_test = test.drop(id_cols + [target_col, pred_col], axis=1)
    y_test, _, results, model = run_model(
            LGBMRegressor(n_estimators=120),
            X_train, y_train, X_val, y_val, X_test,
            metric_names, results,
            early_stopping=True)
    output.append([y_test, model])

model.feature_importances_

results

y_test = np.mean([x[0] for x in output], axis=0)

y_test

from sklearn.metrics import mean_squared_error

mean_squared_error(weeks[-1:][0].duracion_int, y_test)

mean_squared_error(days_data[-1:][0].duracion_int, y_test)
