==> notebooks/Kervy/05-Embeddings.ipynb <==
%load_ext autoreload
%autoreload 2
%matplotlib inline
%config InlineBackend.figure_format = 'retina'

from pandas_summary import DataFrameSummary
from pathlib import Path

import surprise
import os, math
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
import seaborn as sns

pd.set_option('display.max_columns', 1000)
pd.set_option('display.max_rows', 400)
sns.set()
os.chdir('/home/krivas/projects/analysis-project/')

DATA        = Path('data')
RAW         = DATA/'raw'
PROCESSED   = DATA/'processed'

from surprise import SVD, NMF, KNNBasic, evaluate, Reader, Dataset
from surprise.model_selection import train_test_split
from sklearn.metrics import confusion_matrix
from sklearn.metrics import roc_auc_score

# Reading data

csv = pd.read_csv(PROCESSED/'HistoricoJob.csv')

csv.columns

csv_group = csv.groupby(['Id_Job', 'Id_Malla'])['hubo_error'].max().reset_index()

csv_group.head()

csv_group.hubo_error.value_counts()

# Testing

from surprise import Reader, Dataset
reader = Reader(rating_scale=(0, 1))

# The columns must correspond to user id, item id and ratings (in that order).
data = Dataset.load_from_df(csv_group, reader)

data.split(n_folds=5)

%%time
# svd
algo = SVD()
evaluate(algo, data, measures=['RMSE'])

%%time
# nmf
algo = NMF()
evaluate(algo, data, measures=['RMSE'])

# Testing NMF, SVD on train/test

reader = Reader(rating_scale=(0, 1))
data   = Dataset.load_from_df(csv_group, reader)
trainset = data.build_full_trainset()

csv_group.head()

### NMF

nmf = NMF(n_factors=20, n_epochs=50, biased=True, verbose=True)

%%time
nmf.fit(trainset)

%%time
from tqdm import tqdm
train_preds = []
for cIdx, iIdx in tqdm(zip(csv_group.Id_Job, csv_group.Id_Malla), total=len(csv_group)):
    train_preds.append(nmf.predict(cIdx, iIdx)[3])

roc_auc_score(csv_group.hubo_error, train_preds)

confusion_matrix(csv_group.hubo_error, np.round(train_preds))

tn, fp, fn, tp = confusion_matrix(csv_group.hubo_error, np.round(train_preds)).ravel()

tp, fp

### SVD

svd = SVD(n_factors=20, n_epochs=50, biased=True, verbose=True)

%%time
svd.fit(trainset)

%%time
from tqdm import tqdm
train_preds = []
for cIdx, iIdx in tqdm(zip(csv_group.Id_Job, csv_group.Id_Malla), total=len(csv_group)):
    train_preds.append(svd.predict(cIdx, iIdx)[3])

roc_auc_score(csv_group.hubo_error, train_preds)

confusion_matrix(csv_group.hubo_error, np.round(train_preds))

tn, fp, fn, tp = confusion_matrix(csv_group.hubo_error, np.round(train_preds)).ravel()

tp, fp

# Saving weights

from surprise.dump import dump

dump(PROCESSED/'svd_20.dump', algo=svd)

dump(PROCESSED/'nmf_20.dump', algo=nmf)
