==> notebooks/Kervy/08-Best_Params.ipynb <==
%load_ext autoreload
%autoreload 2
%matplotlib inline
%config InlineBackend.figure_format = 'retina'

from pandas_summary import DataFrameSummary
from pathlib import Path

import os, math
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
import seaborn as sns

pd.set_option('display.max_columns', 1000)
pd.set_option('display.max_rows', 400)
sns.set()
os.chdir('/home/krivas/projects/analysis-project/')
from src.utils import convert_int, convert_date,\
                        add_median_features, add_date_diffs, add_days_count, add_date_features,\
                        make_set, run_model, apply_cats


from sklearn.metrics import mean_squared_error

DATA        = Path('data')
RAW         = DATA/'raw'
PROCESSED   = DATA/'processed'

# Reading data

csv = pd.read_csv(PROCESSED/'HistoricoJob.csv')

csv.columns

target_date = int(csv.Fecha_Ejec_Inicio_Int.max())
start_date = 20180401

%%time
days_data = []
for date in range(start_date, target_date + 1):
    if date in csv.Fecha_Ejec_Inicio_Int.unique():
        days_data.append(make_set(csv.loc[csv['Fecha_Ejec_Inicio_Int'] == date], csv.loc[(csv['Fecha_Ejec_Inicio_Int'] == target_date) & (csv['hubo_error'] == 0)]))

# Preprocessing

from tqdm import tqdm

%%time
for temp in tqdm(days_data):
    add_date_diffs(temp, csv)
    add_days_count(temp, csv)
    add_date_features(temp)
    add_median_features(temp, csv, 3)

# Reading dumps
from surprise import dump

svd = dump.load(PROCESSED/'svd_15.dump')[1]

# Adding biases
uid = svd.trainset._raw2inner_id_users
iid = svd.trainset._raw2inner_id_items

for temp in tqdm(days_data):
    # Adding factors features
    fact_job = np.array([svd.pu[uid[cIdx]] for cIdx in temp.Id_Job])
    fact_mesh = np.array([svd.qi[iid[iIdx]] for iIdx in temp.Id_Malla])
    # for ixf in range(fact_job.shape[1]):
    temp['f_job']  = fact_job.prod(axis=1)
    temp['f_mesh']  = fact_mesh.prod(axis=1)
    temp['fail_bias'] = temp.apply(lambda row: svd.predict(row['Id_Job'], row['Id_Malla'])[3], axis=1)
    

days_data[0].head()

# Grid Search

from src.utils import run_model
from lightgbm import LGBMRegressor
from sklearn.model_selection import GridSearchCV, StratifiedKFold
import lightgbm

target_col = 'duracion_int'
pred_col = 'Prediccion'
id_cols = ['Fecha_Ejec_Inicio_Int', 'Id_Job', 'Id_Malla']
metric_names = ['mse']
days_data[len(days_data) - 1][pred_col] = 0

%%time

train, val, test = pd.concat(days_data[:len(days_data)-1], axis=0), days_data[-1], days_data[-1]
X_train, y_train = train.drop(id_cols + [target_col], axis=1), \
                       train[target_col]
X_test = test.drop(id_cols + [target_col, pred_col], axis=1)

estimator = LGBMRegressor(n_jobs=1, random_state=42)

param_grid = {
    'n_estimators': [200, 400],
    'boosting_type': ['dart', 'gbdt'],
    'num_leaves': [31, 50],
    'max_depth': [30, 50]
}

gbm = GridSearchCV(estimator, param_grid, cv=5)

gbm.fit(X_train, y_train)

print('Best parameters found by grid search are:', gbm.best_params_)

gbm.best_params_

# Cross validation with best params

from sklearn.metrics import mean_absolute_error
dftrainLGB = lightgbm.Dataset(data = X_train, label = y_train, feature_name = list(X_train))

params = {'objective': 'regression', 'num_leaves': 50, 'max_depth': 50, 'n_estimators': 400, 'random_state': 42}

cv_results = lightgbm.cv(
        params,
        dftrainLGB,
        num_boost_round=100,
        nfold=3,
        metrics='mae',
        early_stopping_rounds=30,
        stratified=False
        )

cv_results
